pyproject.toml
[project]
name = "trunkroom-data"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "eijun", email = "eijun.tei@gmail.com" }
]
dependencies = [
    "streamlit>=1.45.0",
    "supabase>=2.15.1",
    "python-dotenv>=1.1.0",
    "beautifulsoup4>=4.13.4",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "pandas>=2.2.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/trunkroom_data"]
